Use the WangBrushItem also in AbstractTileFillTool
Allows highlighting the places where Wang Fill didn't find a tile also
for the Bucket Fill and Shape Fill tools.
